[Silver III] Title: 주유소, Time: 900 ms, Memory: 165728 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -8,13 +8,13 @@ public class Main {
 
 		int[] station = new int[n];
 		int[] road = new int[n - 1];
-		int total = 0;
+		long total = 0;
 
 		for (int i = 0; i < n - 1; i++) road[i] = sc.nextInt();
 		for (int i = 0; i < n; i++) station[i] = sc.nextInt();
 		
 		for (int i = 0; i < n - 1; i++) {
-			int dis = 0;
+			long dis = 0;
 			int skip = 0;
 			for(int j = i; j < n - 1; j++) {
 				if(station[i] > station[j]) {
@@ -27,7 +27,7 @@ public class Main {
 			total += station[i] * dis;
 			i += (skip - 1);
 		}
-        
 		System.out.println(total);
 	}
+
 }
